0.7.3

+++ client/delete.php
@@ -8,10 +8,27 @@ session_start();
 $site = new viewsite();
 $client = new client();
 
-$site->starthead();
-$site->backmenu();
-$client->deleteclient();
-$site->endhead();
+if (isset($_SESSION['access']))
+
+{
+    if ($_SESSION['access'] >= 1)
+    {
+        $site->starthead();
+        $site->backmenu();
+        $client->deleteclient();
+        $site->endhead();
+    }
+    else
+    {
+        $site->error();
+    }
+}
+
+else
+
+{
+    $site->error();
+}
 
 
 ob_end_flush();
